requirements.txt
# Core dependencies
gradio==4.44.0
transformers==4.44.2
sentence-transformers==3.0.0
chromadb==0.4.24
pypdf==4.0.0
torch==2.3.0
python-dotenv==1.0.1
accelerate==0.30.0

# Supporting libraries
langchain==0.2.0
langchain-community==0.2.0
huggingface-hub>=0.23.2,<1.0
numpy==1.24.3